[IMROVE] modified customer for account details and improve the customer...

See merge request enigma-camp/gosling-class/batch-1/m-shadiq-firmansyah-basiru/backend/challenge/wmb-api!1

+++ src/main/java/com/enigma/wmb_api/entity/BillDetail.java
@@ -21,11 +21,14 @@ public class BillDetail {
     @JoinColumn(name = "bill_id",nullable = false)
     @JsonBackReference
     private Bill bill;
+
     @ManyToOne
     @JoinColumn(name = "menu_id", nullable = false)
     private Menu menu;
+
     @Column(name = "qty", nullable = false)
     private Integer qty;
+
     @Column(name = "price", nullable = false)
-    private Float price;
+    private Integer price;
 }
